Decouples edges from nodes
