+++ beagle/edges/edge.py
