+++ beagle/nodes/node.py
@@ -1,4 +1,3 @@
-import inspect
 from abc import ABCMeta
 from collections import defaultdict
 from typing import Any, Dict, List, Tuple
